Orders crud

+++ isd-backend/src/main/java/com/example/isdbackend/service/payment/PaymentCalculator.java
@@ -8,7 +8,7 @@ public class PaymentCalculator {
         return userOrder.getPrice() * 0.5 + deliveryPrice;
     }
 
-    protected double paymentCalc2(int deliveryPrice, UserOrderView userOrder, UserOrderView nextUserOrder) {
+    protected double paymentCalc2(double deliveryPrice, UserOrderView userOrder, UserOrderView nextUserOrder) {
         double price = (Math.max(userOrder.getPrice(), nextUserOrder.getPrice()) * 0.5) + Math.min(userOrder.getPrice(), nextUserOrder.getPrice());
 
         return price + deliveryPrice;
